simplify and support all arguments - remove unused dependencies - consolidate code into 1 file - Manually use new ArgParser.allowAnything() instead of command classes to allow any args for either command
