--- lib/src/command/serve_command.dart
@@ -1,100 +1,0 @@
-/*
-Copyright 2018 Workiva Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
----------------------------------------------------------------------
-
-This software or document includes material copied from or derived 
-from webdev https://github.com/dart-lang/webdev
-
-The original license from webdev follows:
-
-Copyright 2017, the Dart project authors. All rights reserved.
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are
-met:
-
-    * Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above
-      copyright notice, this list of conditions and the following
-      disclaimer in the documentation and/or other materials provided
-      with the distribution.
-    * Neither the name of Google Inc. nor the names of its
-      contributors may be used to endorse or promote products derived
-      from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-*/
-import 'dart:async';
-
-import 'command_base.dart';
-
-/// Command to execute pub run build_runner serve.
-class ServeCommand extends CommandBase {
-  @override
-  final name = 'serve';
-
-  @override
-  final description = 'Run a local web development server and a file system'
-      ' watcher that re-builds on changes.';
-
-  @override
-  String get invocation => '${super.invocation} [<directory>[:<port>]]...';
-
-  ServeCommand() : super(releaseDefault: false) {
-    // TODO(nshahan) Expose more args passed to build_runner serve.
-    // build_runner might expose args for use in wrapping scripts like this one.
-    argParser
-      ..addOption('hostname',
-          help: 'Specify the hostname to serve on', defaultsTo: 'localhost')
-      ..addFlag('log-requests',
-          defaultsTo: false,
-          negatable: false,
-          help: 'Enables logging for each request to the server.');
-  }
-
-  @override
-  List<String> getArgs() {
-    var arguments = super.getArgs();
-
-    var hostname = argResults['hostname'] as String;
-    if (hostname != null) {
-      arguments.addAll(['--hostname', hostname]);
-    }
-
-    if (argResults['log-requests'] == true) {
-      arguments.add('--log-requests');
-    }
-
-    // The remaining arguments should be interpreted as [<directory>[:<port>]].
-    arguments.addAll(argResults.rest);
-
-    return arguments;
-  }
-
-  @override
-  Future<int> run() => runCore('serve');
-}
