+++ pubspec.yaml
@@ -15,11 +15,8 @@ environment:
 dependencies:
   args: ">=0.13.7 < 2.0.0"
   io: ^0.3.2+1
-  meta: ^1.1.2
   path: ^1.5.1
-  pub_semver: ^1.3.2
   stack_trace: ^1.9.2
-  yaml: ^2.1.13
 
   # For Dart 2 use build runner
   # build_runner: ^0.8.0
